backend/requirements.txt
fastapi==0.121.2
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.44
python-dotenv==1.0.1
pydantic==2.12.4
python-dateutil==2.9.0.post0
aiofiles==24.1.0